# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coastaq multi-vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  BUYER: "BUYER",
  SELLER: "SELLER",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role: RegisterRequestRole;
  shopName?: string;
  shopDescription?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  BUYER: "BUYER",
  SELLER: "SELLER",
  ADMIN: "ADMIN",
} as const;

export type ShopSubscriptionStatus =
  (typeof ShopSubscriptionStatus)[keyof typeof ShopSubscriptionStatus];

export const ShopSubscriptionStatus = {
  TRIAL: "TRIAL",
  ACTIVE: "ACTIVE",
  EXPIRED: "EXPIRED",
  CANCELLED: "CANCELLED",
} as const;

export interface Shop {
  id: string;
  name: string;
  slug?: string;
  description?: string;
  logo?: string;
  banner?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  website?: string;
  address?: string;
  city?: string;
  country?: string;
  businessHours?: string;
  accentColor?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  tiktokUrl?: string;
  twitterUrl?: string;
  youtubeUrl?: string;
  isApproved: boolean;
  isSuspended?: boolean;
  userId: string;
  subscriptionStatus?: ShopSubscriptionStatus;
  productCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  shop?: Shop;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CreateShopRequest {
  name: string;
  description?: string;
}

export interface UpdateShopRequest {
  name?: string;
  slug?: string;
  description?: string;
  logo?: string;
  banner?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  website?: string;
  address?: string;
  city?: string;
  country?: string;
  businessHours?: string;
  accentColor?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  tiktokUrl?: string;
  twitterUrl?: string;
  youtubeUrl?: string;
}

export type ProductCondition =
  (typeof ProductCondition)[keyof typeof ProductCondition];

export const ProductCondition = {
  NEW: "NEW",
  USED: "USED",
  REFURBISHED: "REFURBISHED",
} as const;

export type _CategoryCount = {
  products?: number;
};

export interface Category {
  id: string;
  name: string;
  parentId?: string;
  children?: Category[];
  _count?: _CategoryCount;
}

export interface Product {
  id: string;
  title: string;
  description?: string;
  price: number;
  stock: number;
  condition: ProductCondition;
  location?: string;
  images: string[];
  categoryId?: string;
  shopId: string;
  shop?: Shop;
  category?: Category;
  createdAt: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateProductRequestCondition =
  (typeof CreateProductRequestCondition)[keyof typeof CreateProductRequestCondition];

export const CreateProductRequestCondition = {
  NEW: "NEW",
  USED: "USED",
  REFURBISHED: "REFURBISHED",
} as const;

export interface CreateProductRequest {
  title: string;
  description?: string;
  price: number;
  stock: number;
  condition: CreateProductRequestCondition;
  location?: string;
  images?: string[];
  categoryId?: string;
}

export type UpdateProductRequestCondition =
  (typeof UpdateProductRequestCondition)[keyof typeof UpdateProductRequestCondition];

export const UpdateProductRequestCondition = {
  NEW: "NEW",
  USED: "USED",
  REFURBISHED: "REFURBISHED",
} as const;

export interface UpdateProductRequest {
  title?: string;
  description?: string;
  price?: number;
  stock?: number;
  condition?: UpdateProductRequestCondition;
  location?: string;
  images?: string[];
  categoryId?: string;
}

export interface CreateCategoryRequest {
  name: string;
  parentId?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PENDING: "PENDING",
  PAID: "PAID",
  SHIPPED: "SHIPPED",
  DELIVERED: "DELIVERED",
  CANCELLED: "CANCELLED",
} as const;

export interface OrderItem {
  id: string;
  orderId: string;
  productId: string;
  product?: Product;
  quantity: number;
  price: number;
}

export interface Order {
  id: string;
  userId: string;
  status: OrderStatus;
  total: number;
  shippingName?: string;
  shippingAddress?: string;
  shippingCity?: string;
  shippingState?: string;
  shippingZip?: string;
  shippingCountry?: string;
  paymentMethod?: string;
  paymentId?: string;
  items?: OrderItem[];
  createdAt: string;
}

export interface CartItem {
  productId: string;
  quantity: number;
  price: number;
  title: string;
  shopId: string;
}

export interface ShippingAddress {
  name: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  country: string;
}

export interface CheckoutRequest {
  items: CartItem[];
  shipping: ShippingAddress;
}

export interface StripeSessionResponse {
  sessionId: string;
  url: string;
}

export interface PaypalOrderResponse {
  orderId: string;
  approvalUrl: string;
}

export interface CapturePaypalRequest {
  paypalOrderId: string;
  orderId: string;
}

export interface Analytics {
  totalUsers: number;
  totalSellers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingApprovals: number;
}

export interface UploadImageRequest {
  /** Base64-encoded image data */
  data: string;
  folder?: string;
}

export interface UploadImageResponse {
  url: string;
  publicId: string;
}

export type ListProductsParams = {
  categoryId?: string;
  search?: string;
  shopId?: string;
  page?: number;
  limit?: number;
};