# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coastaq multi-vendor marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List shops
      responses:
        "200":
          description: Shops list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new shop (max 3 per seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopRequest"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          description: Maximum shop limit reached
...
  /shops/my:
    get:
      operationId: getMyShop
      tags: [shops]
      summary: Get current seller's first shop (backward compat)
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

    put:
      operationId: updateMyShop
      tags: [shops]
      summary: Update current seller's first shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
      responses:
        "200":
          description: Shop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/slug/{slug}:
    get:
      operationId: getShopBySlug
      tags: [shops]
      summary: Get a shop by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Shop not found
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    put:
      operationId: updateShop
      tags: [shops]
      summary: Update a specific shop by ID (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
      responses:
        "200":
          description: Shop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Shop not found
    delete:
      operationId: deleteShop
      tags: [shops]
      summary: Delete a specific shop (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shop deleted
...
components:
  schemas:
...
    Shop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        logo:
          type: string
        banner:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        website:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        businessHours:
          type: string
        accentColor:
          type: string
        facebookUrl:
          type: string
        instagramUrl:
          type: string
        tiktokUrl:
          type: string
        twitterUrl:
          type: string
        youtubeUrl:
          type: string
        isApproved:
          type: boolean
        isSuspended:
          type: boolean
        userId:
          type: string
        subscriptionStatus:
          type: string
          enum: [TRIAL, ACTIVE, EXPIRED, CANCELLED]
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isApproved
        - userId
        - createdAt
    CreateShopRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    UpdateShopRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        logo:
          type: string
        banner:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        website:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        businessHours:
          type: string
        accentColor:
          type: string
        facebookUrl:
          type: string
        instagramUrl:
          type: string
        tiktokUrl:
          type: string
        twitterUrl:
          type: string
        youtubeUrl:
          type: string
...