# lib/api-zod/src/generated/types/createShopRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coastaq multi-vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateShopRequest {
  name: string;
  description?: string;
}